+++ sc-memory/cpp/kpm/sc_agent.hpp
@@ -6,9 +6,10 @@
 
 #pragma once
 
-#include "sc-memory/cpp/sc_object.hpp"
-#include "sc-memory/cpp/sc_memory.hpp"
-#include "sc-memory/cpp/utils/sc_log.hpp"
+#include "../sc_object.hpp"
+#include "../sc_memory.hpp"
+#include "../utils/sc_log.hpp"
+#include "../sc_keynodes.hpp"
 
 #include "../generated/sc_agent.generated.hpp"
 
@@ -45,9 +46,6 @@ public:
   _SC_EXTERN explicit ScAgentAction(ScAddr const & cmdClassAddr, char const * name, sc_uint8 accessLvl = sc_access_lvl_make_max);
   _SC_EXTERN virtual ~ScAgentAction();
 
-  // Just for internal usage
-  static bool InitSets(ScMemoryContext & ctx);
-
 protected:
   _SC_EXTERN virtual sc_result Run(ScAddr const & listenAddr, ScAddr const & edgeAddr, ScAddr const & otherAddr) override;
 
@@ -64,9 +62,7 @@ public:
   static _SC_EXTERN ScAddr const & GetCommandInitiatedAddr();
   static _SC_EXTERN ScAddr const & GetCommandFinishedAddr();
   static _SC_EXTERN ScAddr const & GetNrelResultAddr();
-  static _SC_EXTERN ScAddr const & GetResultCodeAddr(sc_result resCode);
-  static _SC_EXTERN sc_result GetResultCodeByAddr(ScAddr const & resultClassAddr);
-
+ 
   /* Create instance of specified command with arguments
    * cmdClassAddr - node that designate command class
    * For each parameter from params would be added attribute rrel_<i + 1>, where
@@ -92,62 +88,17 @@ public:
    * process isn't finished. It could be finished with that result.
    */
   static _SC_EXTERN sc_result GetCommandResultCode(ScMemoryContext & ctx, ScAddr const & cmdAddr);
+  
+  /* Same as GetCommandResultCode, but returns ScAddr of result code set.
+   * If it can't be determined, then return empty value
+   */
+  static _SC_EXTERN ScAddr GetCommandResultCodeAddr(ScMemoryContext & ctx, ScAddr const & cmdAddr);
 
 private:
   ScAddr m_cmdClassAddr;
 
 private:
-  SC_PROPERTY(Keynode("command_initiated"), ForceCreate(ScType::NodeConstClass))
-  static ScAddr ms_commandInitiatedAddr;
-
-  SC_PROPERTY(Keynode("command_in_progress"), ForceCreate(ScType::NodeConstClass))
-  static ScAddr ms_commandProgressdAddr;
-
-  SC_PROPERTY(Keynode("command_finished"), ForceCreate(ScType::NodeConstClass))
-  static ScAddr ms_commandFinishedAddr;
-
-  SC_PROPERTY(Keynode("nrel_result"), ForceCreate(ScType::NodeConstNoRole))
-  static ScAddr ms_nrelResult;
-
-  SC_PROPERTY(Keynode("nrel_common_template"), ForceCreate(ScType::NodeConstNoRole))
-  static ScAddr ms_nrelCommonTemplate;
-
-  // result codes
-  SC_PROPERTY(Keynode("sc_result"), ForceCreate(ScType::NodeConstClass))
-  static ScAddr ms_keynodeScResult;
-
-  SC_PROPERTY(Keynode("sc_result_ok"), ForceCreate(ScType::NodeConstClass))
-  static ScAddr ms_keynodeScResultOk;
-
-  SC_PROPERTY(Keynode("sc_result_no"), ForceCreate(ScType::NodeConstClass))
-  static ScAddr ms_keynodeScResultNo;
-
-  SC_PROPERTY(Keynode("sc_result_unknown"), ForceCreate(ScType::NodeConstClass))
-  static ScAddr ms_keynodeScResultUnknown;
-
-  SC_PROPERTY(Keynode("sc_result_error"), ForceCreate(ScType::NodeConstClass))
-  static ScAddr ms_keynodeScResultError;
-
-  SC_PROPERTY(Keynode("sc_result_error_invalid_params"), ForceCreate(ScType::NodeConstClass))
-  static ScAddr ms_keynodeScResultErrorInvalidParams;
-
-  SC_PROPERTY(Keynode("sc_result_error_invalid_type"), ForceCreate(ScType::NodeConstClass))
-  static ScAddr ms_keynodeScResultErrorInvalidType;
-
-  SC_PROPERTY(Keynode("sc_result_error_io"), ForceCreate(ScType::NodeConstClass))
-  static ScAddr ms_keynodeScResultErrorIO;
-
-  SC_PROPERTY(Keynode("sc_result_invalid_state"), ForceCreate(ScType::NodeConstClass))
-  static ScAddr ms_keynodeScResultInvalidState;
-
-  SC_PROPERTY(Keynode("sc_result_error_not_found"), ForceCreate(ScType::NodeConstClass))
-  static ScAddr ms_keynodeScResultErrorNotFound;
-
-  SC_PROPERTY(Keynode("sc_result_error_no_write_rights"), ForceCreate(ScType::NodeConstClass))
-  static ScAddr ms_keynodeScResultErrorNoWriteRights;
-
-  SC_PROPERTY(Keynode("sc_result_error_no_read_rights"), ForceCreate(ScType::NodeConstClass))
-  static ScAddr ms_keynodeScResultErrorNoReadRights;
+  
 };
 
 #define AGENT_NAME_TYPE(__Name__) __Name__##Type